// mytradeplatform-backend/src/payment/dto/create-payment-method.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsString, IsNotEmpty } from 'class-validator';

/**
 * Defines the data structure for creating a new payment method.
 */
export class CreatePaymentMethodDto {
  /**
   * The PaymentMethod ID generated by Stripe on the client-side (e.g., 'pm_card_visa').
   * @example 'pm_card_visa'
   */
  @ApiProperty({
    description: "The PaymentMethod ID generated by Stripe on the client-side (e.g., 'pm_card_visa').",
    example: 'pm_card_visa',
  })
  @IsString()
  @IsNotEmpty()
  stripePaymentMethodId: string;
}